[ToDoList][James] feature:add should_return_status_not_found_when_delete_company_given_not_exist_id test

+++ src/main/java/com/oocl/todolist/service/ToDoService.java
@@ -29,7 +29,9 @@ public class ToDoService {
         return toDoRepository.save(oldToDo);
     }
     //ToDo
-    public void deleteById(int id) {
+    public void deleteById(int id) throws NoSuchIdException {
+        if(!toDoRepository.findById(id).isPresent())
+            throw new NoSuchIdException();
         toDoRepository.deleteById(id);
     }
 }
